Fixes #22.  Force Event model to find users by lowercase email, since devise will automatically downcase emails on User creation.
This comes into play when adding a user to an event who already has an account.  Their email address will be in the DB lowercase because devise forced that.  We run find_or_create_by(email: xxxx) which will not find it because of the uppercase/lowercase mismatch.  Then, it will drop into the create block and spin up an object for that email address.  Then, when it tries to save the model, it will fail, because devise downcases it again and it fails the uniqueness constraint.  Thus, we have an unpersisted object which would otherwise correspond to a DB row and this wreaks havok on the RSVP creation step.