[Fixes#40] Adds Maximum Attendance
Adds a (nullable) attribute to events for the maximum number of guests
that the event can support.  Also adds model methods to support this
functionality. If a user tries to RSVP after the max is reached,
they won't have the form available to actually RSVP.  Additionally,
if the event has a max set, we tell invitees, so there's some pressure
to respond.

This commit also refactors a bunch of view logic, pushing some of it
down into the model, otherwise, just moving around if/elses so they're
a bit more compact.