A User Collects Sipping Points
----- Sipping Points -----

Gamified components (Event, RSVP) will inherently have a SP value, but a
user only counts those points towards themself if the relevant event has
tipped.

The current logic weights towards attending events, but the logic is
centralized into the SippingPoints module and the #sipping_points and
(unfortunately) the #attendance_points methods on User, Event, and RSVP.

Tweaking should be straightforward, but make sure you update the tests
in test/models/user_test.rb if you do tweak the math.

----- is_tipped bit -----

In order to massively simplify work when calculating a user's sipping
points, a boolean column was added to Event to indicate whether the
event has tipped or not . This allows simple joins and reduces the
number of queries required on read, at the expense of add'l DB writes.
Specifically, an additional DB write will happen when an event tips.

Keeping this updated should happen automatically via callbacks on
Event and Rsvp but for safety a rake task exists to ensure DB integrity.
Although the DB should never get out of sync, it's possible if somebody
is hacking around via raw SQL. To fix this situation, the
new task, sipping_point:update_event_tipped_column, will crawl every
event and update based on its associated Rsvp records. For now, that's a
quick process, but if this is necessary long-term, throw that in the
background.

As a nice follow-up, it would be nice to use the new scope on event,
coming_rsvps, to DRY up some code.